refactor: split handler bindings into package
